Fix calls to qsort_arg
Turns out these comparison functions are not only passed to qsort but also
called. In this case the changed operand order really does matter.
This was causing the MIPS postgres on FreeBSD to try to dereference an
ASCII char buffer instead of a valid pointer...

Avoid these errors by renaming the comparator functions and adding two macros
QSORT_ARG_COMPARATOR_PTR and CALL_QSORT_ARG_COMPARATOR that may or may not
switch the argument order depending on FreeBSD vs Linux