Add _Alignas to the va_data member for external data
CHERI128 was crashing with an alignment error when storing to this
member so hopefully this fixes it. On x86 all tests still pass with
this change.
I initially added the _Alignas to all the variable length char arrays in
this file but that broke lots of stuff. For this struct it seems safe.