Fix function ABI for oldstyle function
It expects all parameters and the return value to be of type Datum when
calling it. Before this change n was being passed in $c3 but the function
used $c3 as the text* so it would crash.

This should work but really the better solution is to just not allow old
style functions for CHERI.